use int instead of long
long can be >32 bits, int can not.
And we don't run on platforms where int is less than 32 bits.

+++ source/fix/fix.h
@@ -81,7 +81,7 @@ COPYRIGHT 1993-1998 PARALLAX SOFTWARE CORPORATION.  ALL RIGHTS RESERVED.
 
 #include "types.h"
 
-typedef long fix;				//16 bits int, 16 bits frac
+typedef int fix;				//16 bits int, 16 bits frac
 typedef short fixang;		//angles
 
 typedef struct quad {
